requirements.txt
anthropic>=0.40.0
httpx==0.28.1
numpy==2.2.6
pandas==2.3.2
pandas-ta>=0.3.14b
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests>=2.31.0
tzdata==2025.2
websockets>=12.0
